___________INSTALL_GOOD

+++ sys/install-lib.ps1
@@ -1,6 +1,17 @@
 param (
-    [string]$serverName
+    [string]$serverName = 'default'
 )
+
+function Set-KeyboardLayouts {
+    $langlist = New-WinUserLanguageList en-US
+    $langlist[0].InputMethodTips.Clear()
+    $langlist[0].InputMethodTips.Add('0409:00000409')
+    $langlist.Add((New-WinUserLanguageList ru-RU)[0])
+    $langlist[1].InputMethodTips.Clear()
+    $langlist[1].InputMethodTips.Add('0419:00000419')
+    Set-WinUserLanguageList $langlist -Force
+    Set-WinUILanguageOverride -Language en-US
+}
 
 #currents
 $scriptDir = Split-Path -Parent $MyInvocation.MyCommand.Path
@@ -11,6 +22,24 @@ Set-Location -Path $scriptDir
 $password = $server.clone.clonePassword
 $user=$server.clone.cloneUser
 $serverIp = $server.clone.cloneServerIp
+
+function Test {
+    $client = New-Object System.Net.Sockets.TcpClient
+    try {
+        $async = $client.BeginConnect($serverIp, 5985, $null, $null)
+        $wait = $async.AsyncWaitHandle.WaitOne(4000, $false) # 4000 ms = 3 seconds timeout
+        if ($wait -and $client.Connected) {
+            $client.EndConnect($async)
+            return $true
+        } else {
+            return $false
+        }
+    } catch {
+        return $false
+    } finally {
+        $client.Close()
+    }
+}
 
 function sharpRdp {
     $programPath = Join-Path $scriptDir "../rdp/SharpRdp.exe"
@@ -44,49 +73,132 @@ function Invoke-RemoteCommand {
 
     return $jobResult
 }
+function WaitForLocalTag {
+    param (
+        [string]$tag,
+        [int]$timeout
+    )
+
+    Write-Host "Waiting for local tag $tag ..."
+    Set-Content -Path 'C:\tag_local_r.txt' -Value $tag
+    $filePath = "C:\tag_local.txt"
+    $startTime = Get-Date
+
+    function IsTag() {
+        if (Test-Path $filePath) {
+            $content = Get-Content -Path $filePath -Raw                            
+            $co = $content -like "*$tag*"
+            if ($co -eq $false) {
+                return 1
+            }
+            $co = $content -like "*timeout*"
+            if ($co) {
+                return -1
+            } else {
+                return 0
+            }
+        } else {
+            return 1
+        }
+    }
+
+    while ($true) {
+        $elapsed = (Get-Date) - $startTime
+        if ($elapsed.TotalSeconds -ge $timeout) {
+            Write-Host "Timeout reached after $timeout seconds."
+            return -1
+        }
+
+        $result = IsTag
+        if ($result -eq 0 -or $result -eq -1) {
+            Write-Host "Tag local '$tag' detected!"
+            return $result
+        }
+        Start-Sleep -Seconds 1
+    }
+}
+
 
 function WaitForTag {
     param (
-        [string]$tag
+        [string]$tag,
+        [int]$timeout
     )
-   while ($true) {
-       Start-Sleep -Seconds 1
-       try {
-           Invoke-RemoteCommand -ScriptBlock {
+    $result = 0
+    $startTime = Get-Date
+
+    while ($true) {
+        Start-Sleep -Seconds 1
+
+        $elapsed = (Get-Date) - $startTime
+        if ($elapsed.TotalSeconds -ge $timeout) {
+            Write-Host "Timeout reached before connection to server."
+            return -1
+        }
+
+        try {
+            $tested = Test
+            if ($tested -eq $false)
+            {
+                Start-Sleep -Seconds 1
+                continue
+            } 
+
+           $result = Invoke-RemoteCommand -ScriptBlock {
                 param (
-                    [string]$tag
+                    [string]$tag,
+                    [int]$timeout
                 )
                 Write-Host "Waiting for tag $tag ..."
                 Set-Content -Path 'C:\tagR.txt' -Value $tag
-                $filePath = "C:\tag1.txt"
+                $filePath = "C:\tag.txt"
+                $startTime = Get-Date
+
                 function IsTag() {
                     if (Test-Path $filePath) {
                         $content = Get-Content -Path $filePath -Raw
-                        $co = $content  -like "*$tag*"
-                        return $co
+                        $co = $content -like "*$tag*"
+                        if ($co -eq $false) {
+                            return 1
+                        }
+                        $co = $content -like "*timeout*"
+                        if ($co) {
+                            return -1
+                        } else {
+                            return 0
+                        }
                     } else {
-                        return $false
+                        return 1
                     }
                 }
+
                 while ($true) {
-                    if (IsTag) {
+                    $elapsed = (Get-Date) - $startTime
+                    if ($elapsed.TotalSeconds -ge $timeout) {
+                        Write-Host "Timeout reached waiting for tag '$tag'."
+                        return -1
+                    }
+
+                    $result = IsTag
+                    if ($result -eq 0 -or $result -eq -1) {
                         Write-Host "Tag '$tag' detected!"
-                        break
+                        return $result
                     }
                     Start-Sleep -Seconds 3
                 }
-        
-            } -Arguments @($tag)   
-           break
-       }
-       catch {
-          Write-Host $_
-          Start-Sleep -Seconds 3
-       }
-       Start-Sleep -Seconds 1
-   }
-   Start-Sleep -Seconds 1
-   Write-Host "Tag found"
+
+            } -Arguments @($tag, $timeout)
+
+            break
+        }
+        catch {
+            Write-Host $_
+            Start-Sleep -Seconds 3
+        }
+
+        Start-Sleep -Seconds 1
+    }
+    return $result
 }
 
 function WaitRestart {
@@ -95,6 +207,16 @@ function WaitRestart {
     )
     Write-Host "Restarting.."
    while ($true) {
+        $tested = Test
+        if ($tested -eq $false)
+        {
+            if ($once)
+            {
+                return
+            }
+            Start-Sleep -Seconds 1    
+            continue
+        }
        try {
            Invoke-RemoteCommand -ScriptBlock { shutdown /r /t 0 /f }
            Start-Sleep -Seconds 3
@@ -113,13 +235,15 @@ function WaitRestart {
 
    while ($true) 
    {
-        if (Test-Connection -ComputerName $serverIp -Count 1 -Quiet) {
+        $tested = Test
+        if ($tested -eq $true)
+        {
             break
         }
         Write-Host "Ping attempt.."
-        Start-Sleep -Seconds 3
+        Start-Sleep -Seconds 2
     }
-    Start-Sleep -Seconds 3
+    Start-Sleep -Seconds 1
    
     while ($true) {
        Start-Sleep -Seconds 1
